requirements.txt
numpy
python-binance
ta-lib
websocket_client